Moved AddMessageQueue and AddMessageBus to Messaging ns
